+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/InputMediaVideo.kt
@@ -20,7 +20,7 @@ data class InputMediaVideo(
     @Optional
     override val height: Int? = null,
     @Optional
-    override val duration: Int? = null,
+    override val duration: Long? = null,
     @Transient
     override val thumb: InputFile? = null
 ) : InputMedia, SizedInputMedia, DuratedInputMedia, ThumbedInputMedia, CaptionedInputMedia, MediaGroupMemberInputMedia {
